Create WEBPAGE1
Most people consider Github just a repository for code and a place where developers can collaborate on projects and visitors can download cool new open-source applications. It's definitely that, but there's much more available to you as a Github user if you know how to take advantage of it. Here's how you can use Github to host your webapp and make it accessible to visitors, as a good-looking website that lists your projects, or even a personal nameplate site you can direct people to.